pyproject.toml
[project]
name = "goit-pythonweb-hw-12"
version = "0.1.0"
description = ""
authors = [
    {name = "RomanTuras",email = "salden.com@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "sqlalchemy (>=2.0.39,<3.0.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "greenlet (>=3.1.1,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "libgravatar (>=1.0.4,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "aiosmtplib (>=4.0.0,<5.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "pydantic[email] (>=2.11.1,<3.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "bcrypt (==4.0.1)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "starlette (>=0.46.1,<0.47.0)",
    "mypy (>=1.15.0,<2.0.0)",
    "cloudinary (>=1.43.0,<2.0.0)",
    "ruff (>=0.11.2,<0.12.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{include = "src"}]
[tool.poetry.group.dev.dependencies]
sphinx = "^8.2.3"

